refactor: View Controller와 ApiController 분리
