+++ src/test/java/cart/controller/AdminApiControllerTest.java
@@ -12,7 +12,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class AdminControllerTest {
+class AdminApiControllerTest {
 
     @LocalServerPort
     int port;
